Split tests into 4 files
