+++ test/src/Test.files
@@ -6,3 +6,4 @@ T_SRCS := \
   state_machine/dummy.test.cpp \
   utils/config.test.cpp \
   utils/vectors/vector_properties.test.cpp \
+   utils/vectors/vector_constructor.test.cpp \
